+++ lib/Controls/Controller.js
@@ -1593,6 +1593,7 @@ class ControlsController extends CoreBase {
 
 	/**
 	 * Create a control
+	 * Danger: This will not delete an existing control from the specified location
 	 * @param {import('../Resources/Util.js').ControlLocation} location Location to place in the grid
 	 * @param {string} newType The type of the new control to create (if any)
 	 * @returns {string | null} controlId
