+++ lib/Data/ImportExport.js
@@ -33,33 +33,6 @@ import { visitEventOptions } from '../Resources/EventDefinitions.js'
 import { compareExportedInstances } from '../Shared/Import.js'
 
 /**
- * Default buttons on fresh pages
- */
-const default_nav_buttons_definitions = [
-	{
-		type: 'pageup',
-		location: {
-			column: 0,
-			row: 0,
-		},
-	},
-	{
-		type: 'pagenum',
-		location: {
-			column: 0,
-			row: 1,
-		},
-	},
-	{
-		type: 'pagedown',
-		location: {
-			column: 0,
-			row: 2,
-		},
-	},
-]
-
-/**
  * @param {import('winston').Logger} logger
  * @param {import("express").Response} res
  * @param {import("express").NextFunction} next
@@ -504,26 +477,6 @@ class DataImportExport extends CoreBase {
 	}
 
 	/**
-	 *
-	 * @param {number} pageCount
-	 * @returns {void}
-	 */
-	createInitialPageButtons(pageCount) {
-		for (let page = 1; page <= pageCount; page++) {
-			for (const definition of default_nav_buttons_definitions) {
-				const location = {
-					...definition.location,
-					pageNumber: page,
-				}
-				const oldControlId = this.page.getControlIdAt(location)
-				if (oldControlId) this.controls.deleteControl(oldControlId)
-
-				this.controls.createButtonControl(location, definition.type)
-			}
-		}
-	}
-
-	/**
 	 * Setup a new socket client's events
 	 * @param {import('../UI/Handler.js').ClientSocket} client - the client socket
 	 * @access public
@@ -692,113 +645,6 @@ class DataImportExport extends CoreBase {
 					style: controlObj.type,
 				})
 				return !!res?.style ? res?.asDataUrl ?? null : null
-			}
-		)
-
-		client.onPromise(
-			'loadsave:reset-page-clear',
-			/**
-			 * @param {number} pageNumber
-			 * @returns {'ok'}
-			 */
-			(pageNumber) => {
-				this.logger.silly(`Reset page ${pageNumber}`)
-
-				// Delete the controls, and allow them to redraw
-				const controlIds = this.page.getAllControlIdsOnPage(pageNumber)
-				for (const controlId of controlIds) {
-					this.controls.deleteControl(controlId)
-				}
-
-				// Clear the references on the page
-				this.page.resetPage(pageNumber)
-
-				// Re-add the nav buttons
-				for (const { type, location } of default_nav_buttons_definitions) {
-					this.controls.createButtonControl(
-						{
-							pageNumber,
-							...location,
-						},
-						type
-					)
-				}
-
-				return 'ok'
-			}
-		)
-
-		client.onPromise(
-			'loadsave:reset-page-nav',
-			/**
-			 * @param {number} pageNumber
-			 * @returns {'ok'}
-			 */
-			(pageNumber) => {
-				// make magical page buttons!
-				for (const { type, location } of default_nav_buttons_definitions) {
-					this.controls.createButtonControl(
-						{
-							pageNumber,
-							...location,
-						},
-						type
-					)
-				}
-
-				return 'ok'
-			}
-		)
-
-		client.onPromise(
-			'loadsave:delete-page',
-			/**
-			 * @param {number} pageNumber
-			 * @returns {'ok' | 'fail'}
-			 */
-			(pageNumber) => {
-				this.logger.silly(`Delete page ${pageNumber}`)
-
-				if (this.page.getPageCount() === 1) return 'fail'
-
-				// Delete the controls, and allow them to redraw
-				const controlIds = this.page.getAllControlIdsOnPage(pageNumber)
-				for (const controlId of controlIds) {
-					this.controls.deleteControl(controlId)
-				}
-
-				// Delete the page
-				this.page.deletePage(pageNumber)
-
-				return 'ok'
-			}
-		)
-
-		client.onPromise(
-			'loadsave:insert-page',
-			/**
-			 * @param {number} asPageNumber
-			 * @returns {'ok' | 'fail'}
-			 */
-			(asPageNumber) => {
-				this.logger.silly(`Insert new page ${asPageNumber}`)
-
-				// Delete the page
-				const pageId = this.page.insertPage(asPageNumber)
-				if (!pageId) throw new Error(`Failed to insert page`)
-
-				// Add nav buttons
-				for (const { type, location } of default_nav_buttons_definitions) {
-					this.controls.createButtonControl(
-						{
-							pageNumber: asPageNumber,
-							...location,
-						},
-						type
-					)
-				}
-
-				return 'ok'
 			}
 		)
 
@@ -1043,15 +889,7 @@ class DataImportExport extends CoreBase {
 			// Reset page 1
 			this.page.resetPage(1) // Note: controls were already deleted above
 			if (!skipNavButtons) {
-				for (const { type, location } of default_nav_buttons_definitions) {
-					this.controls.createButtonControl(
-						{
-							pageNumber: 1,
-							...location,
-						},
-						type
-					)
-				}
+				this.page.createPageDefaultNavButtons(1)
 			}
 
 			// Delete other pages
