+++ lib/Page/Controller.js
@@ -3,6 +3,7 @@ import CoreBase from '../Core/Base.js'
 import { oldBankIndexToXY } from '../Shared/ControlId.js'
 import { nanoid } from 'nanoid'
 import jsonPatch from 'fast-json-patch'
+import { default_nav_buttons_definitions } from './Defaults.js'
 
 const PagesRoom = 'pages'
 
@@ -98,6 +99,89 @@ class PageController extends CoreBase {
 		client.onPromise('pages:unsubscribe', () => {
 			client.leave(PagesRoom)
 		})
+
+		client.onPromise(
+			'pages:delete-page',
+			/**
+			 * @param {number} pageNumber
+			 * @returns {'ok' | 'fail'}
+			 */
+			(pageNumber) => {
+				this.logger.silly(`Delete page ${pageNumber}`)
+
+				if (this.getPageCount() === 1) return 'fail'
+
+				// Delete the controls, and allow them to redraw
+				const controlIds = this.getAllControlIdsOnPage(pageNumber)
+				for (const controlId of controlIds) {
+					this.controls.deleteControl(controlId)
+				}
+
+				// Delete the page
+				this.deletePage(pageNumber)
+
+				return 'ok'
+			}
+		)
+
+		client.onPromise(
+			'pages:insert-page',
+			/**
+			 * @param {number} asPageNumber
+			 * @returns {'ok' | 'fail'}
+			 */
+			(asPageNumber) => {
+				this.logger.silly(`Insert new page ${asPageNumber}`)
+
+				// Delete the page
+				const pageId = this.insertPage(asPageNumber)
+				if (!pageId) throw new Error(`Failed to insert page`)
+
+				// Add nav buttons
+				this.createPageDefaultNavButtons(asPageNumber)
+
+				return 'ok'
+			}
+		)
+
+		client.onPromise(
+			'pages:reset-page-clear',
+			/**
+			 * @param {number} pageNumber
+			 * @returns {'ok'}
+			 */
+			(pageNumber) => {
+				this.logger.silly(`Reset page ${pageNumber}`)
+
+				// Delete the controls, and allow them to redraw
+				const controlIds = this.getAllControlIdsOnPage(pageNumber)
+				for (const controlId of controlIds) {
+					this.controls.deleteControl(controlId)
+				}
+
+				// Clear the references on the page
+				this.resetPage(pageNumber)
+
+				// Re-add the nav buttons
+				this.createPageDefaultNavButtons(pageNumber)
+
+				return 'ok'
+			}
+		)
+
+		client.onPromise(
+			'pages:reset-page-nav',
+			/**
+			 * @param {number} pageNumber
+			 * @returns {'ok'}
+			 */
+			(pageNumber) => {
+				// make magical page buttons!
+				this.createPageDefaultNavButtons(pageNumber)
+
+				return 'ok'
+			}
+		)
 	}
 
 	/**
@@ -388,7 +472,7 @@ class PageController extends CoreBase {
 	 * Reset a page to defaults and empty
 	 * Note: Controls will be orphaned if not explicitly deleted by the caller
 	 * @param {number} pageNumber - the page id
-	 * @param {boolean} [redraw = false] - <code>true</code> if the graphics should invalidate
+	 * @param {boolean} [redraw = true] - <code>true</code> if the graphics should invalidate
 	 * @returns ControlIds referenced on the page
 	 * @access public
 	 */
@@ -413,6 +497,24 @@ class PageController extends CoreBase {
 		this.#commitChanges([pageNumber], redraw)
 
 		return removedControls
+	}
+
+	/**
+	 * Recreate the default nav buttons on a page
+	 * @param {number} pageNumber - the page id
+	 * @access public
+	 */
+	createPageDefaultNavButtons(pageNumber) {
+		for (const { type, location } of default_nav_buttons_definitions) {
+			const fullLocation = {
+				...location,
+				pageNumber,
+			}
+			const oldControlId = this.page.getControlIdAt(fullLocation)
+			if (oldControlId) this.controls.deleteControl(oldControlId)
+
+			this.controls.createButtonControl(fullLocation, type)
+		}
 	}
 
 	/**
@@ -598,7 +700,7 @@ class PageController extends CoreBase {
 
 			setImmediate(() => {
 				// @ts-ignore
-				this.registry.data.importExport.createInitialPageButtons(1)
+				this.createPageDefaultNavButtons(1)
 			})
 		}
