+++ lib/Page/Defaults.js
@@ -1,0 +1,26 @@
+/**
+ * Default buttons on fresh pages
+ */
+export const default_nav_buttons_definitions = [
+	{
+		type: 'pageup',
+		location: {
+			column: 0,
+			row: 0,
+		},
+	},
+	{
+		type: 'pagenum',
+		location: {
+			column: 0,
+			row: 1,
+		},
+	},
+	{
+		type: 'pagedown',
+		location: {
+			column: 0,
+			row: 2,
+		},
+	},
+]
