+++ webui/src/Buttons/Pages.tsx
@@ -1,11 +1,12 @@
 import React, { useCallback, useContext, useRef } from 'react'
 import { CButton, CButtonGroup, CCol, CRow } from '@coreui/react'
-import { PagesContext, SocketContext, socketEmitPromise } from '../util'
+import { PagesContext, SocketContext, socketEmitPromise } from '../util.js'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { faPlus, faTrash } from '@fortawesome/free-solid-svg-icons'
-import { PageModel } from '@companion/shared/Model/PageModel'
-import { GenericConfirmModal, GenericConfirmModalRef } from '../Components/GenericConfirmModal'
-import { usePageCount } from '../Hooks/usePagesInfoSubscription'
+import { faPlus, faTrash, faPencil } from '@fortawesome/free-solid-svg-icons'
+import { PageModel } from '@companion/shared/Model/PageModel.js'
+import { GenericConfirmModal, GenericConfirmModalRef } from '../Components/GenericConfirmModal.js'
+import { usePageCount } from '../Hooks/usePagesInfoSubscription.js'
+import { EditPagePropertiesModal, EditPagePropertiesModalRef } from './EditPageProperties.js'
 
 interface PagesListProps {
 	setPageNumber: (page: number) => void
@@ -18,6 +19,7 @@ export function PagesList({ setPageNumber }: PagesListProps): JSX.Element {
 	const pageCount = usePageCount()
 
 	const deleteRef = useRef<GenericConfirmModalRef>(null)
+	const editRef = useRef<EditPagePropertiesModalRef>(null)
 
 	const goToPage = useCallback(
 		(e: React.MouseEvent<HTMLButtonElement>) => {
@@ -30,10 +32,19 @@ export function PagesList({ setPageNumber }: PagesListProps): JSX.Element {
 		[setPageNumber]
 	)
 
+	const configurePage = useCallback((e: React.MouseEvent<HTMLButtonElement>) => {
+		const pageNumber = Number(e.currentTarget.getAttribute('data-page'))
+		const pageInfoRaw = e.currentTarget.getAttribute('data-page-info')
+		const pageInfo = pageInfoRaw ? JSON.parse(pageInfoRaw) : null
+		if (!isNaN(pageNumber) && pageInfo) {
+			editRef.current?.show(pageNumber, pageInfo)
+		}
+	}, [])
+
 	const doInsertPage = useCallback((e: React.MouseEvent<HTMLButtonElement>) => {
 		const pageNumber = Number(e.currentTarget.getAttribute('data-page'))
 		if (!isNaN(pageNumber)) {
-			socketEmitPromise(socket, 'loadsave:insert-page', [pageNumber]).catch((e) => {
+			socketEmitPromise(socket, 'pages:insert-page', [pageNumber]).catch((e) => {
 				console.error('Page insert failed', e)
 			})
 		}
@@ -53,7 +64,7 @@ export function PagesList({ setPageNumber }: PagesListProps): JSX.Element {
 			],
 			'Delete',
 			() => {
-				socketEmitPromise(socket, 'loadsave:delete-page', [pageNumber]).catch((e) => {
+				socketEmitPromise(socket, 'pages:delete-page', [pageNumber]).catch((e) => {
 					console.error('Page delete failed', e)
 				})
 			}
@@ -64,6 +75,7 @@ export function PagesList({ setPageNumber }: PagesListProps): JSX.Element {
 		<CRow>
 			<CCol xs={12}>
 				<GenericConfirmModal ref={deleteRef} />
+				<EditPagePropertiesModal ref={editRef} />
 
 				<table className="table table-responsive-sm">
 					<thead>
@@ -81,6 +93,7 @@ export function PagesList({ setPageNumber }: PagesListProps): JSX.Element {
 								info={info}
 								pageCount={pageCount}
 								goToPage={goToPage}
+								configurePage={configurePage}
 								doInsertPage={doInsertPage}
 								doDeletePage={doDeletePage}
 							/>
@@ -88,7 +101,7 @@ export function PagesList({ setPageNumber }: PagesListProps): JSX.Element {
 						<tr>
 							<td></td>
 							<td></td>
-							<td>
+							<td style={{ textAlign: 'center' }}>
 								<CButtonGroup>
 									<CButton
 										color="warning"
@@ -114,24 +127,36 @@ interface PageListRowProps {
 	info: PageModel | undefined
 	pageCount: number
 	goToPage: (e: React.MouseEvent<HTMLButtonElement>) => void
+	configurePage: (e: React.MouseEvent<HTMLButtonElement>) => void
 	doInsertPage: (e: React.MouseEvent<HTMLButtonElement>) => void
 	doDeletePage: (e: React.MouseEvent<HTMLButtonElement>) => void
 }
 
-function PageListRow({ id, info, pageCount, goToPage, doInsertPage, doDeletePage }: PageListRowProps) {
+function PageListRow({ id, info, pageCount, goToPage, configurePage, doInsertPage, doDeletePage }: PageListRowProps) {
 	return (
 		<tr>
-			<td>
+			<td style={{ width: 0 }}>
 				<CButton color="primary" variant="ghost" onClick={goToPage} data-page={id}>
 					{id}
 				</CButton>
 			</td>
 			<td>{info?.name ?? ''}</td>
-			<td>
+			<td style={{ width: 0 }}>
 				<CButtonGroup>
+					<CButton
+						color="info"
+						size="sm"
+						onClick={configurePage}
+						title="Edit page"
+						data-page={id}
+						data-page-info={JSON.stringify(info)}
+					>
+						<FontAwesomeIcon icon={faPencil} />
+					</CButton>
 					<CButton color="warning" size="sm" onClick={doInsertPage} title="Insert page before" data-page={id}>
 						<FontAwesomeIcon icon={faPlus} />
 					</CButton>
+
 					<CButton
 						color="danger"
 						size="sm"
